create datafeeds module
